+++ use-cases/cyclones/workflows/meta.yml
@@ -1,4 +1,4 @@
 # Configuration file for use case metadata
 # Load with OmegaConf
 
-root: ./use-cases/tensorflow/mnist
+root: ./use-cases/cyclones
