--- use-cases/mnist/tensorflow/env-files/preproc-env.yml
@@ -1,10 +1,0 @@
-name: preproc-env
-channels:
-  - tensorflow
-  - conda-forge
-dependencies:
-  - tensorflow
-  - python=3.9.12
-  - pip:
-      - tfx
-      - scipy==1.9.3
