--- use-cases/mnist/tensorflow/testing-workflow.yml
@@ -1,26 +1,0 @@
-# Workflow definition in classical style
-# Load with OmegaConf
-# Other configuration files to merge with this file via OmegaConf
-
-conf-dependencies:
-  - meta.yml
-
-steps:
-  - tfx_test:
-        doc: Subpipeline via TFX for training
-        command: python ${root}/test.py
-        env:
-          file: ${root}/env-files/preproc-env.yml
-          prefix: ${root}/.venv-preproc
-        args:
-          output: test_pipe
-          models: ${root}/pipelines/train_pipe/Trainer/model/
-          dataset: ${root}/data/test/test.tfrecords
-
-preprocessEnvironment:
-  class: Directory
-  path: .venv-preproc
-
-preprocessScript:
-  class: File
-  path: train.py
