--- use-cases/mnist/tensorflow/training-workflow.yml
@@ -1,34 +1,0 @@
-# Workflow definition in classical style
-# Load with OmegaConf
-# Other configuration files to merge with this file via OmegaConf
-
-conf-dependencies:
-  - meta.yml
-
-steps:
-  - dataset:
-      doc: Preparation of the MNIST dataset into TFRecord
-      command: python ${root}/preproc.py
-      env:
-        file: ${root}/env-files/preproc-env.yml
-        prefix: ${root}/.venv-preproc
-      args:
-        directory: ${root}/data
-  - tfx_train:
-      doc: Subpipeline via TFX for training
-      command: python ${root}/train.py
-      env: 
-        file: ${root}/env-files/preproc-env.yml
-        prefix: ${root}/.venv-preproc
-      args:
-        name: train_pipe
-        datadir: ${root}/data/train
-        component: ${root}/components.py
-
-preprocessEnvironment:
-  class: Directory
-  path: .venv-preproc
-
-preprocessScript:
-  class: File
-  path: train.py
