--- use-cases/mnist/torch/README.md
@@ -1,3 +1,0 @@
-# MNIST: toy example for DT workflows
-
-Read the latest version of the docs [here](https://intertwin-eu.github.io/T6.5-AI-and-ML/docs/use-cases/mnist.html).
