--- use-cases/mnist/torch/env-files/preproc-env.yml
@@ -1,8 +1,0 @@
-name: preproc-env
-channels:
-  - pytorch
-  - conda-forge
-dependencies:
-  - python=3.9.12
-  - torchvision
-  - pytorch::cpuonly
