--- use-cases/tensorflow/cyclones/env-files/tensorflow-env.yml
@@ -1,14 +1,0 @@
-# TODO: Install itwinai
-
-name: tensorflow-env
-channels:
-  - tensorflow
-  - conda-forge
-dependencies:
-  - tensorflow
-  - python=3.9.12
-  - pip:
-      - tfx
-      - scipy==1.9.3
-      - scikit-learn
-      - gdown
