--- use-cases/tensorflow/cyclones/workflows/meta.yml
@@ -1,4 +1,0 @@
-# Configuration file for use case metadata
-# Load with OmegaConf
-
-root: ./use-cases/tensorflow/cyclones
