--- use-cases/tensorflow/cyclones/workflows/workflow-train.yml
@@ -1,25 +1,0 @@
-# Workflow definition in classical style
-# Load with OmegaConf
-# Other configuration files to merge with this file via OmegaConf
-conf-dependencies:
-  - meta.yml
-
-steps:
-  - DataLoading:
-      doc: Get data for the training from Google Drive
-      command: python ${root}/dataloader.py
-      env:
-        file: ${root}/env-files/tensorflow-env.yml
-        prefix: ${root}/.venv-tensorflow
-      args:
-        output_dir: ${root}/data
-  - Training:
-      doc: Run Tensorflow Cyclone training
-      command: python -u ${root}/train.py
-      env:
-        file: ${root}/env-files/tensorflow-env.yml
-        prefix: ${root}/.venv-tensorflow
-      args:
-        batch_size: 512
-        epochs: 3
-        root_dir: ${root}
